requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
APScheduler==3.10.4
pytz==2023.3
python-dotenv==1.0.0